Create Build a pile of cubes (6kyu)
The code returns number of cubes (n) which are used to make a pile of cubes of total volume of m. The cube at the bottom has volume of (a^3) and each of the next cubes has volume of (a-1)^3, the next (a-2)^3 ... and the last one has volume of 1^3.